Refactoring, skeleton 'selection' popup
