+++ src/pages/knowledgegraph/SelectionDetails/index.ts
@@ -1,0 +1,1 @@
+export * from './SelectionDetails';
